started V - g conversion function and added DAS read program
Started the function to convert from voltage to acceleration. Inserted
all the calibration, except we still need the charge amp conversion
(which is listed as a vague 10-100 mV/pC.

Also added a new program to read data from the HP DAS. Right now it
successfully reads and plots the current trace. Needs a lot of work
still, but it's basically functional.